render token rows on server side
