+++ app/components/TokenTable.tsx
@@ -1,8 +1,9 @@
 "use client";
 
 import React, { useState, useEffect } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
 import { FungibleToken } from "@/app/types";
+
+import { DynamicTokenRow } from "@/app/components";
 
 interface TokenTableProps {
   source: string;
@@ -20,8 +21,6 @@ const TokenTable = ({
   const [currentPage, setCurrentPage] = useState(1);
   const [sortedTokens, setSortedTokens] = useState<FungibleToken[]>([]);
   const itemsPerPage = perPage || 8; // Adjust the number of items per page as needed
-  const router = useRouter();
-  const searchParams = useSearchParams();
 
   useEffect(() => {
     // Sort tokens by total value in descending order
@@ -94,80 +93,13 @@ const TokenTable = ({
           </thead>
 
           <tbody className="divide-y divide-white/5">
-            {currentTokens.map((token) => {
-              const tokenImage = token.content.links.image;
-              const tokenSymbol =
-                token.token_info.symbol ||
-                token.content.metadata.symbol ||
-                token.id;
-              const tokenBalance = (
-                token.token_info.balance /
-                Math.pow(10, token.token_info.decimals)
-              ).toFixed(5);
-              const tokenPrice =
-                token.token_info.price_info?.price_per_token || "N/A";
-              const tokenValue =
-                token.token_info.price_info?.total_price?.toFixed(2) || "N/A";
-
-              return (
-                <tr
-                  key={token.id}
-                  onClick={() =>
-                    router.push(
-                      `/portfolio/${walletAddress}?${searchParams.toString()}&tokenDetails=${
-                        token.id
-                      }`,
-                    )
-                  }
-                  className="group hover:cursor-pointer"
-                >
-                  {/* ICON */}
-                  <td className="  py-4 pl-4  sm:pl-6 lg:pl-7">
-                    <div className="flex items-center gap-x-4">
-                      {tokenImage ? (
-                        <img
-                          src={tokenImage}
-                          alt="Token Icon"
-                          className="h-12 w-12 rounded-full bg-gray-800 ring-1 ring-white ring-opacity-0 ring-offset-1 transition-all duration-200 ease-in-out group-hover:ring-opacity-100"
-                        />
-                      ) : (
-                        <div className="skeleton h-12 w-12 shrink-0 rounded-full" />
-                      )}
-                    </div>
-                  </td>
-
-                  {/* SYMBOL */}
-                  <td className=" py-4 pl-0 pr-4  sm:pr-8">
-                    <div className="flex gap-x-3">
-                      <div className="rounded-md bg-gray-700/40 px-3 py-1.5 text-sm font-medium text-gray-400 ring-1 ring-inset ring-white/10 transition-all duration-200 ease-in-out group-hover:bg-gray-700/60 group-hover:text-white">
-                        {tokenSymbol}
-                      </div>
-                    </div>
-                  </td>
-
-                  {/* BALANCE */}
-                  <td className="py-4 pl-0 pr-4 text-sm leading-6 sm:pr-8 lg:pr-20">
-                    <div className="flex items-center justify-end gap-x-2 sm:justify-start">
-                      <div className="text-gray-400 group-hover:text-white">
-                        {tokenBalance}
-                      </div>
-                    </div>
-                  </td>
-
-                  {/* PRICE */}
-                  <td className=" py-4 pl-0 pr-8 text-sm leading-6 text-gray-400 transition-all duration-200 ease-in-out group-hover:text-white  lg:pr-20">
-                    {typeof tokenPrice === "number"
-                      ? `$${tokenPrice}`
-                      : tokenPrice}
-                  </td>
-
-                  {/* VALUE */}
-                  <td className=" py-4 pl-0 pr-4 text-right text-sm leading-6 text-gray-400 transition-all duration-200 ease-in-out group-hover:text-white  sm:pr-6 lg:pr-8">
-                    {tokenValue}
-                  </td>
-                </tr>
-              );
-            })}
+            {currentTokens.map((token) => (
+              <DynamicTokenRow
+                key={token.id}
+                token={token}
+                walletAddress={walletAddress}
+              />
+            ))}
           </tbody>
         </table>
       </div>
